Version 1.0

+++ ciapi-ejb/src/java/repository/CategoryFacadeLocal.java
@@ -4,6 +4,12 @@ import entity.Category;
 import java.util.List;
 import javax.ejb.Local;
 
+/**
+ * Локальный интерфейс, задающий операции над объектами типа Category 
+ * - создание, редактирование, удаление, поиск по первичному ключу, 
+ * выборка всех сущностей и другие.
+ * @author Теплякова Е.А.
+ */
 @Local
 public interface CategoryFacadeLocal {
 
@@ -19,6 +25,6 @@ public interface CategoryFacadeLocal {
 
     List<Category> findRange(int[] range);
 
-    int count();
+    long count();
     
 }
