+++ ciapi-ejb/src/java/repository/PhotoFacadeLocal.java
@@ -4,6 +4,12 @@ import entity.Photo;
 import java.util.List;
 import javax.ejb.Local;
 
+/**
+ * Локальный интерфейс, задающий операции над объектами 
+ * типа Photo - создание, редактирование, удаление, 
+ * поиск по первичному ключу, выборка всех сущностей и другие.
+ * @author Теплякова Е.А.
+ */
 @Local
 public interface PhotoFacadeLocal {
 
@@ -19,6 +25,6 @@ public interface PhotoFacadeLocal {
 
     List<Photo> findRange(int[] range);
 
-    int count();
+    long count();
     
 }
