+++ ciapi-ejb/src/java/repository/StatusFacadeLocal.java
@@ -4,6 +4,12 @@ import entity.Status;
 import java.util.List;
 import javax.ejb.Local;
 
+/**
+ * Локальный интерфейс, задающий операции над объектами 
+ * типа Status - создание, редактирование, удаление, 
+ * поиск по первичному ключу, выборка всех сущностей и другие.
+ * @author Теплякова Е.А.
+ */
 @Local
 public interface StatusFacadeLocal {
 
@@ -19,6 +25,6 @@ public interface StatusFacadeLocal {
 
     List<Status> findRange(int[] range);
 
-    int count();
+    long count();
     
 }
